[Simple Bank] - Fixed account id from int32 to int64

+++ 6-simple-bank/db/sqlc/entry.sql.go
@@ -40,7 +40,7 @@ WHERE id = $1
 LIMIT 1
 `
 
-func (q *Queries) GetEntry(ctx context.Context, id int32) (Entry, error) {
+func (q *Queries) GetEntry(ctx context.Context, id int64) (Entry, error) {
 	row := q.db.QueryRowContext(ctx, getEntry, id)
 	var i Entry
 	err := row.Scan(
