+++ 6-simple-bank/db/sqlc/models.go
@@ -9,7 +9,7 @@ import (
 )
 
 type Account struct {
-	ID        int32     `json:"id"`
+	ID        int64     `json:"id"`
 	Owner     string    `json:"owner"`
 	Balance   float64   `json:"balance"`
 	Currency  string    `json:"currency"`
@@ -17,7 +17,7 @@ type Account struct {
 }
 
 type Entry struct {
-	ID        int32 `json:"id"`
+	ID        int64 `json:"id"`
 	AccountID int64 `json:"account_id"`
 	// can be negative or positive
 	Amount    float64   `json:"amount"`
@@ -25,7 +25,7 @@ type Entry struct {
 }
 
 type Transfer struct {
-	ID            int32 `json:"id"`
+	ID            int64 `json:"id"`
 	FromAccountID int64 `json:"from_account_id"`
 	ToAccountID   int64 `json:"to_account_id"`
 	// must be positive
